feat(vita,user): Make Adrenaline Application config smaller
- Reduce the size of options
- Remove unused fields
- New config header magic number 2 for the new configuration version

Reduce from 60 bytes to 32 bytes

+++ user/menu.h
@@ -85,7 +85,7 @@ typedef struct {
   int type;
   uint32_t color;
   int (* callback)();
-  int *value;
+  uint8_t *value;
   char **options;
   int n_options;
 } MenuEntry;
